19 feb 2020

+++ StudentsServicceProjectNetBeans/src/main/java/bd/edu/seu/models/Student.java
@@ -5,7 +5,7 @@ public class Student {
     private String name;
     private double cgpa;
 
-    public Student(int id, String name, double cgpa) {
+    public Student(long id, String name, double cgpa) {
         this.id = id;
         this.name = name;
         this.cgpa = cgpa;
